+++ Libplanet.Extensions.Cocona/Commands/ApvCommand.cs
@@ -12,6 +12,7 @@ namespace Libplanet.Extensions.Cocona.Commands
     using Libplanet.Crypto;
     using Libplanet.KeyStore;
     using Libplanet.Net;
+    using Libplanet.Net.Transports;
 
     public class ApvCommand
     {
@@ -274,6 +275,39 @@ namespace Libplanet.Extensions.Cocona.Commands
             {
                 Utils.PrintTable(("Field", "Value"), data);
             }
+        }
+
+        [Command(Description = "Query app protocol version (a.k.a. APV) of target node.")]
+        public void Query(
+            [Argument(
+                Name = "TARGET",
+#pragma warning disable MEN002 // Line is too long
+                Description = "Comma seperated peer information({pubkey},{host},{port}) of target node. (e.g. 027bd36895d68681290e570692ad3736750ceaab37be402442ffb203967f98f7b6,9c.planetarium.dev,31236)")]
+#pragma warning restore MEN002 // Line is too long
+            string peerInfo)
+        {
+            BoundPeer peer;
+            AppProtocolVersion apv;
+
+            try
+            {
+                peer = BoundPeer.ParsePeer(peerInfo);
+            }
+            catch
+            {
+                throw Utils.Error($"Failed to parse peer. Please check the input. [{peerInfo}]");
+            }
+
+            try
+            {
+                apv = peer.QueryAppProtocolVersion();
+            }
+            catch
+            {
+                throw Utils.Error($"Failed to query app protocol version.");
+            }
+
+            Console.WriteLine(apv.Token);
         }
 
         private AppProtocolVersion ParseAppProtocolVersionToken(string? token)