+++ Libplanet/Net/BoundPeer.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Diagnostics.Contracts;
+using System.Globalization;
 using System.Net;
 using System.Runtime.Serialization;
 using Libplanet.Crypto;
@@ -52,6 +53,60 @@ namespace Libplanet.Net
 
         public static bool operator !=(BoundPeer left, BoundPeer right) =>
             Operator.Weave(left, right);
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BoundPeer"/> class from
+        /// comma-separated string.</summary>
+        /// <param name="peerInfo">A comma-separated string have format {pubkey},{host},{port}.
+        /// </param>
+        /// <returns>A <see cref="BoundPeer"/> from given data.</returns>
+        /// <exception cref="ArgumentException">Thrown when the given peerInfo is invalid.
+        /// </exception>
+        public static BoundPeer ParsePeer(string peerInfo)
+        {
+            if (peerInfo is null)
+            {
+                throw new ArgumentNullException(nameof(peerInfo));
+            }
+
+            string[] tokens = peerInfo.Split(',');
+            if (tokens.Length != 3)
+            {
+                throw new ArgumentException(
+                    $"'{peerInfo}', should have format <pubkey>,<host>,<port>",
+                    nameof(peerInfo)
+                );
+            }
+
+            if (!(tokens[0].Length == 130 || tokens[0].Length == 66))
+            {
+                throw new ArgumentException(
+                    $"'{peerInfo}', a length of public key must be 130 or 66 in hexadecimal," +
+                    $" but the length of given public key '{tokens[0]}' doesn't.",
+                    nameof(peerInfo)
+                );
+            }
+
+            try
+            {
+                var pubKey = new PublicKey(ByteUtil.ParseHex(tokens[0]));
+                var host = tokens[1];
+                var port = int.Parse(tokens[2], CultureInfo.InvariantCulture);
+
+                // FIXME: It might be better to make Peer.AppProtocolVersion property nullable...
+                return new BoundPeer(
+                    pubKey,
+                    new DnsEndPoint(host, port));
+            }
+            catch (Exception e)
+            {
+                throw new ArgumentException(
+                    $"{nameof(peerInfo)} seems invalid. [{peerInfo}]",
+                    nameof(peerInfo),
+                    innerException: e
+                );
+            }
+        }
 
         /// <inheritdoc/>
         public override void GetObjectData(